Separate registration handling from MessageHandler
